requirements.txt
# Google Cloud & Vertex AI
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.14.0
vertexai>=1.38.0

# Google ADK & GenAI
google-adk>=0.1.0
google-genai>=0.3.0

# LangGraph & LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-vertexai>=0.0.6

# Frontend
streamlit>=1.28.0

# Embedding (Gemini API)
google-generativeai>=0.8.0

# PDF e chunking
pypdf>=4.0.0

# Utils
python-dotenv>=1.0.0

# Testes
pytest>=7.0.0
pytest-cov>=4.0.0
